# src/main/resources/static/api/v2/spec/satakunta-events.swagger.yaml
swagger: '2.0'
info:
  title: Satakunta Events information API
  description: >-
    Satakunta Events provides categorized data on events and places using JSON-LD
    format.
    
    
    
    Events can be searched by date and location.

    
    JSON-LD format is streamlined using include mechanism. API users can request
    that certain

    fields are included directly into the result, instead of being hyperlinks to
    objects.


    Several fields are multilingual. These are implemented as object with each
    language variant

    as property. In this specification each multilingual field has (fi,sv,en)
    property triplet as

    example.
  termsOfService: (depends on data provider)
  version: v2
host: www.satakuntaevents.fi
basePath: /api/v2
schemes:
  - https
  - http
produces:
  - application/json
...
parameters:
...
  event_sort_param:
    name: sort
    in: query
    type: string
    description: >-
      Sort the returned events in the given order. Possible sorting criteria are
      'start_time', 'end_time', 'duration' and 'last_modified_time'. The default
      ordering is '-last_modified_time'.
...
  include_param:
    name: include
    in: query
    type: array
    items:
      type: string
    description: >-
      Embed given reference-type fields directly into the response, otherwise
      they are returned as URI references.
  event_text_param:
    name: text
    in: query
    description: >-
      Search (case insensitive) through all multilingual text fields (name,
      description, short_description, info_url) of an event (every language).
      Multilingual fields contain the text that users are expected to care
      about, thus multilinguality is useful discriminator.
    required: false
    type: string
  event_last_modified_param:
    name: last_modified_since
    in: query
    description: Search for events that have been modified since or at this time.
    type: string
    format: dateTime
  event_bbox_param:
    name: bbox
    in: query
    type: array
    items:
      type: string
    minItems: 4
    maxItems: 4
    description: >-
      Search for events that are within this bounding box. Decimal coordinates
      are given in order west, south, east, north. Period is used as decimal
      separator.
  event_start_param:
    name: start
    in: query
    type: string
    format: date-time
    description: >-
      Search for events beginning or ending after this time. Dates can be
      specified using ISO 8601 ("2016-01-12") and additionally "today".
  event_end_param:
    name: end
    in: query
    type: string
    format: date-time
    description: >-
      Search for events beginning or ending before this time. Dates can be
      specified using ISO 8601 ("2016-01-12") and additionally "today".
  event_data_source_param:
    name: data_source
    in: query
    type: string
    description: Search for events that come from the specified source system
  event_location_param:
    name: location
    in: query
    type: array
    items:
      type: integer
    description: >-
      Search for events in given locations as specified by id. Multiple ids are
      separated by comma
  event_keyword_param:
    name: keyword
    in: query
    type: string
    description: >-
      Search for events with given keywords as specified by id. Multiple ids are
      separated by comma
  event_recurring_param:
    name: recurring
    in: query
    type: string
    enum:
      - super
      - sub
    description: >-
      Search for events based on whether they are part of recurring event set.
      'super' specifies recurring, while 'sub' is non-recurring.
  event_min_duration_param:
    name: min_duration
    in: query
    type: integer
    description: Search for events that are longer than given time in seconds
  event_max_duration_param:
    name: max_duration
    in: query
    type: integer
    description: Search for events that are shorter than given time in seconds
  event_organization_param:
    name: publisher
    in: query
    type: string
    description: Search for events published by the given organization
definitions:
  meta_definition:
    title: Meta
    description: >-
      Meta record for result pagination. All results from API are paginated, ie.
      delivered in chunks of X results. This records describes how many results
      there are in total, and how to access previous and next pages.
    type: object
    properties:
      count:
        description: Total amount of items found
        type: integer
      next:
        description: URL for the next page of items
        type: string
        format: url
      previous:
        description: URL for the previous page of items
        type: string
        format: url
...
paths:
...
  /event/:
    get:
      summary: Return a list of events
      description: >-
        <h2 id="filtering-retrieved-events">Filtering retrieved events</h2>

        <p>Query parameters can be used to filter the retrieved events by

        the following criteria.</p>

        <h3 id="event-time">Event time</h3>

        <p>Use <code>start</code> and <code>end</code> to restrict the date
        range of returned events.

        Any events that intersect with the given date range will be
        returned.</p>

        <p>The parameters <code>start</code> and <code>end</code> can be given
        in the following formats:</p>

        <ul>

        <li>ISO 8601 (including the time of day)</li>

        <li>yyyy-mm-dd</li>

        </ul>

        <p>In addition, <code>today</code> can be used as the value.</p>

        <p>Example:</p>

        <pre><code>event/?start=2014-01-15&amp;end=2014-01-20

        </code></pre>

        <p><a href="?start=2014-01-15&amp;end=2014-01-20" title="json">See the
        result</a></p>

        <h3 id="event-location">Event location</h3>

        <h4 id="bounding-box">Bounding box</h4>

        <p>To restrict the retrieved events to a geographical region, use

        the query parameter <code>bbox</code> in the format</p>

        <pre><code>bbox=west,south,east,north

        </code></pre>

        <p>Where <code>west</code> is the longitude of the rectangle's western
        boundary,

        <code>south</code> is the latitude of the rectangle's southern boundary,

        and so on.</p>

        <p>Example:</p>

        <pre><code>event/?bbox=21.7267856118,61.4623491897,21.8666842696,61.4623491897

        </code></pre>

        <p><a href="?bbox=21.7267856118,61.4623491897,21.8666842696,61.4623491897" title="json">See the
        result</a></p>

        <h4 id="specific-location">Specific location</h4>

        <p>To restrict the retrieved events to a known location(s), use

        the query parameter <code>location</code>, separating values by commas

        if you wish to query for several locations.</p>

        <p>Location ids are found at the <code>place</code> endpoint, which
        lists the

        locations in decreasing number of events found.

        <p>Example:</p>

        <pre><code>event/?location=pori:4

        </code></pre>

        <p><a href="?location=pori:4" title="json">See the result</a></p>

        <h3 id="event-category">Event category</h3>

        <p>To restrict the retrieved events by category, use the query

        parameter <code>keyword</code>, separating values by commas if you wish
        to

        query for several locations.</p>

        <p>Keyword ids are found at the <code>keyword</code> endpoint, which
        lists the

        keywords in decreasing number of events found. The common keywords

        used in all events originate from keywords added to Satakunta Events system.</p>

        <p>Example:</p>

        <pre><code>event/?keyword=pori:topic:fairs

        </code></pre>

        <p><a href="?keyword=pori:topic:fairs" title="json">See the result</a></p>

        <h3 id="event-text">Event text</h3>

        <p>To find out events that contain a specific string in any of the text

        fields, use the query parameter <code>text</code>.</p>

        <p>Example:</p>

        <pre><code>event/?text=jazz

        </code></pre>

        <p><a href="?text=jazz" title="json">See the result</a></p>

        <h2 id="getting-detailed-data">Getting detailed data</h2>

        <p>In the default case, keywords, locations, and other fields that

        refer to separate resources are only displayed as simple references.</p>

        <p>If you want to include the complete data from related resources in

        the current response, use the keyword <code>include</code>. For
        example:</p>

        <pre><code>event/?include=location,keywords

        </code></pre>

        <p><a href="?include=location,keywords" title="json">See the
        result</a></p>

        <h2 id="ordering">Ordering</h2>

        <p>Default ordering is decreasing order by
        <code>last_modified_time</code>.

        You may also order results by <code>start_time</code>,
        <code>end_time</code>,

        <code>name</code> and <code>duration</code>. For example:</p>

        <pre><code>event/?sort=end_time

        </code></pre>

        <p><a href="?sort=name" title="json">See the result</a></p>
      operationId: Event_list
      tags:
        - event
      responses:
        '200':
          description: List of events
          schema:
            type: object
            properties:
              meta:
                $ref: '#/definitions/meta_definition'
              data:
                type: array
                items:
                  $ref: '#/definitions/event'
      parameters:
        - $ref: '#/parameters/include_param'
        - $ref: '#/parameters/event_text_param'
        - $ref: '#/parameters/event_last_modified_param'
        - $ref: '#/parameters/event_start_param'
        - $ref: '#/parameters/event_end_param'
        - $ref: '#/parameters/event_bbox_param'
        - $ref: '#/parameters/event_data_source_param'
        - $ref: '#/parameters/event_location_param'
        - name: division
          in: query
          type: string
          description: >-
            You may filter places by specific OCD division id, or by division
            name. The latter query checks all divisions with the name,
            regardless of division type. NOTE: Not in use in Satakunta Events API
        - $ref: '#/parameters/event_keyword_param'
        - $ref: '#/parameters/event_recurring_param'
        - $ref: '#/parameters/event_min_duration_param'
        - $ref: '#/parameters/event_max_duration_param'
        - $ref: '#/parameters/event_organization_param'
        - $ref: '#/parameters/event_sort_param'
    post:
      summary: Create a new event
      operationId: event_create
      tags:
        - event
      responses:
        '201':
          description: >-
            Object has been succesfully created. A copy of the object is
            returned in response body and headers contain Location pointing to
            the created event
          schema:
            $ref: '#/definitions/event'
        '400':
          description: >-
            Input format was not correct, eg. mandatory field was missing or
            JSON was malformed
        '401':
          description: User was not authenticated
        '403':
          description: User does not have necessary permissions
      parameters:
        - name: event_object
          in: body
          schema:
            $ref: '#/definitions/event'
  '/event/{id}/':
    get:
      summary: Retrieve single event by id
      operationId: event_retrieve
      tags:
        - event
      responses:
        '200':
          description: Single event object
          schema:
            $ref: '#/definitions/event'
      parameters:
        - name: id
          in: path
          type: string
          description: Event identifier as defined in event schema
          required: true
    put:
      summary: Update an event
      description: >-
        Events can be updated if the user has appropriate access permissions.
        The original implementation behaves like PATCH, ie. if some field is
        left out from the PUT call, its value is retained in database. In order
        to ensure consistent behaviour, users should always supply every field
        in PUT call.
      operationId: event_update
      tags:
        - event
      responses:
        '200':
          description: Event has been succesfully replaced
          schema:
            $ref: '#/definitions/event'
        '400':
          description: >-
            Input format was not correct, eg. mandatory field was missing or
            JSON was malformed
        '401':
          description: User was not authenticated
        '403':
          description: User does not have necessary permissions
      parameters:
        - type: string
          in: path
          name: id
          required: true
          description: Identifier for the event to be updated
        - name: event_object
          in: body
          description: >-
            Event object that should replace the existing event, note that some
            implementations may retain unspecified fields at their original
            values.
          schema:
            $ref: '#/definitions/event'